idle animation
you may notice that the art is a lot better for these frames, and also that there are 13 of them. That's because after staring at my perfect boy for so long, I had lost sight of his beauty, but we've reconnected and at some point I'm gonna go back and redo like... most of the sprites I've done so far. Yep.